server/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# ML/AI
torch>=2.1.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# ControlNet (опционально)
controlnet-aux>=0.0.7

# Изображения
Pillow>=10.2.0

# Утилиты
pydantic>=2.5.0
